Resolves #49 Use BEST_EFFORT to switch controller
As reported in Issue https://github.com/RethinkRobotics/baxter_simulator/issues/49
**ros_control upgrade breaks baxter_gazebo simulation**

A pull request to ros_controls enforced STRICT controller switching to
mean that already stopped controllers cannot be stopped again:
https://github.com/ros-controls/ros_control/pull/200

This resulted in baxter_simulator being unable to start any controllers
since it preemptively closes controllers whether or not they have been
started. This was caused by the chosen strategy for switching controllers.

Passing in a value of STRICT to the controller manager didn't matter
before because it was never properly implemented. According to the
service Request definition:

  * STRICT means that switching will fail if anything goes wrong (an invalid
    controller name, a controller that failed to start, etc. )
  * BEST_EFFORT means that even when something goes wrong with on controller,
    the service will still try to start/stop the remaining controllers

We were relying on STRICT functionality not being fully implemented.
Instead, we will use BEST_EFFORT for a more forgiving controller
switching strategy. This design decision should be revisited in the
development cycle.