Added Sim Started Pub to notify when Loaded
This commit adds a new Latched publisher on
/robot/sim/started topic, which will publish
an std_msgs::Empty message when the simulator's
controllers have fully loaded. This is useful
when launching the baxter_simulator in conjunction
with a ROS node that should only be executed when
the simulator is finished loading, eg.

    # Wait for the All Clear for emulator starting
    import rospy
    from std_msgs.msg import Empty
    rospy.init_node("foobar")
    rospy.wait_for_message("/robot/sim/started", Empty)

This snippit will block until the simulator has
finished loading.