Moves block collision origin for Pick and Place
The IK Pick and Place demo's block was constantly rolling,
making it difficult to grasp repeatably. This commit moves
the Collision and Visual origin to be the same as the Inertial
origin. This prevents the block from interacting with the
table properly, frequently rolling off the table.

Additionally, this commit changes the block pose default to
spawn with respect to the world frame, as the /base frame
of baxter will not necessarily exist when "roslaunch" invokes
both the simulator startup and the IK pick and place demo.