Re #72 Added joint name lists to Pose Kinematics
To avoid the bad assumption that any non-fixed joint with
"left" or "right" in its name is an arm joint, this commit
adds a difinitive list of arm joints to the parameter server
under robot_config/<side>_config/joint_names, just like on
the real robot. This way only the ros_parameter is consulted
for a list of joints to care about, since it is known at the
time of the node's launch.