Added conditional compile for Saucy yaml-cpp
To get the ros_buildfarm running again, yaml-cpp
is checked in baxter_sim_controllers CMakeLists.txt to see
if it is at least version 0.5. If it is an earlier version
(as is the case in the EOL Ubuntu Saucy), then grippers
commands are disabled. This will allow the buildfarm to
compile on Saucy, but only function on Trusty and later.